refactor(cli): replace custom CLI framework with cobra commands
Remove hand-rolled flag parsing, usage formatting, and version handling
New structure:
- main.go: root cobra.Command, version, init
- daemon.go: daemon flags and orchestration via cobra
- commands.go: IPC subcommands via factory pattern
- model.go: model command tree (list/set/delete)
